Fix findbugs violations in openflowjava
- Invocation of toString on an array

- Check for sign of bitwise operation:
Example: ((event.detail & SWT.SELECTED) > 0)

Using bit arithmetic and then comparing with the greater than operator
can lead to unexpected results (of course depending on the value of
SWT.SELECTED). If SWT.SELECTED is a negative number, this is a candidate
for a bug. Even when SWT.SELECTED is not negative, it seems good practice
to use '!= 0' instead of '> 0'

- Reliance on default encoding - performs a byte to String (or String to byte)
conversion, and will assume that the default platform encoding is suitable.

- Primitive value is boxed then unboxed to perform primitive coercion

- Unread field

- Result of integer multiplication cast to long - possible overflow

- Dead store to local variable - assigns a value to a local variable, but
the value is not read or used in any subsequent instruction.

- Method uses the same code for two switch clauses

- Load of known null value

- Consider returning a zero length array rather than null

- Redundant nullcheck of value known to be non-null

- Field not initialized in constructor but dereferenced without null check

- Non-transient non-serializable instance field in serializable class

- Class implements same interface as superclass

+++ openflowjava/openflow-protocol-impl/src/main/java/org/opendaylight/openflowjava/protocol/impl/core/connection/AbstractStackedOutboundQueue.java
@@ -68,7 +68,7 @@ abstract class AbstractStackedOutboundQueue implements OutboundQueue {
 
         for (int i = unflushedSegments.size(); i <= segmentOffset; ++i) {
             final StackedSegment newSegment = StackedSegment.create(first.getBaseXid()
-                    + StackedSegment.SEGMENT_SIZE * i);
+                    + StackedSegment.SEGMENT_SIZE * (long)i);
             LOG.debug("Adding segment {}", newSegment);
             unflushedSegments.add(newSegment);
         }
